agents/sota_sdk/pyproject.toml
[project]
name = "sota-sdk"
version = "0.3.0"
description = "SOTA Agent SDK — Build and deploy marketplace agents in minutes"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "web3>=7.0.0",
    "eth-account>=0.13.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "aiohttp>=3.10.0",
    "websockets>=12.0",
    "pydantic>=2.9.0",
    "uvicorn>=0.32.0",
    "fastapi>=0.115.0",
]

[project.optional-dependencies]
cost = ["paid-python>=1.0.5"]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "black>=24.0.0",
    "ruff>=0.6.0",
]

[project.scripts]
sota = "sota_sdk.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
